+++ tests/test_profile.py
@@ -2,7 +2,7 @@
 
 from bs4 import BeautifulSoup
 
-from britishcycling_clubs import club_profile_url
+from britishcycling_clubs import profile_url
 from britishcycling_clubs.profile import (
     _club_name_from_profile,
     _total_members_from_profile,
@@ -11,9 +11,7 @@ from britishcycling_clubs.profile import (
 
 def test_club_profile_url__happy_path() -> None:
     """Test that correct URL is returned."""
-    assert (
-        club_profile_url("000") == "https://www.britishcycling.org.uk/club/profile/000/"
-    )
+    assert profile_url("000") == "https://www.britishcycling.org.uk/club/profile/000/"
 
 
 # Partial extract from actual page